% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labdat.r
\name{ccCleanLabLoad}
\alias{ccCleanLabLoad}
\title{TryCatch function for ccCleanLabLoadDirect}
\usage{
ccCleanLabLoad(vsFieldCodes = "", lDbCon = list(), nDelay = 20, nTimes = 6)
}
\arguments{
\item{lDbCon}{Database credentials}

\item{nDelay}{Delay in seconds between each try}

\item{nTimes}{Number of times to repeat (tries)}

\item{vsIdSampeKobo}{vsIdSampeKobo Field codes to extract; vector of field codes. Names must be the same as in CleanLab column "field_code" (e.g. "PREZ 123")}

\item{sDbCleanLabVersion}{Version of DbCleanLab to use;  Default: "DbCleanLab". "DbCleanLabDevelopment" should be used for testing during development}
}
\value{
list of dataframes
}
\description{
TryCatch function for ccCleanLabLoadDirect
}
\author{
Daniel Wächter, \email{daniel.waechter@bfh.ch}, Thorsten Behrens, \email{thorsten.behrens@bfh.ch} and Marie Hertzog \email{marie.hertzog@bfh.ch}
}
