% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labdat.r
\name{ccCleanLabTransform}
\alias{ccCleanLabTransform}
\title{Extract, combine and clean data from Soildat}
\usage{
ccCleanLabTransform(
  lCleanLab,
  sUseCase = "CaseDebug",
  sFnDictsXls,
  sConfigFile = "config.yml"
)
}
\arguments{
\item{sUseCase}{Specific use case as defined in the main name encoding dict}

\item{lDicts}{List of KOBO naming and convention dictionaries}

\item{lDbCon}{List of database credentials generated with Creds}

\item{sDbSoildat}{Name of the database containing Soildat data}

\item{sDbCleanlab}{Name of the atabase containing clean lab data}

\item{sProject}{Project to extract; only one project can be extracted}
}
\value{
list of dataframes
}
\description{
Load various datasets from Soildat based on project names
}
\author{
Daniel Wächter, \email{daniel.waechter@bfh.ch}, Thorsten Behrens, \email{thorsten.behrens@bfh.ch} and Marie Hertzog \email{marie.hertzog@bfh.ch}
}
