% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labdat.r
\name{ccQueryMeasurement}
\alias{ccQueryMeasurement}
\title{Builds a postgres query for selecting soil layer by observation id from Soildat}
\usage{
ccQueryMeasurement(vsFieldCodes)
}
\arguments{
\item{sObsIdsQuery}{String of soildat observation ids}
}
\value{
string with postgres query for soildat
}
\description{
Builds a postgres query for selecting soil layer by observation id from Soildat
}
\author{
Daniel Wächter, \email{daniel.waechter@bfh.ch} and Thorsten Behrens, \email{thorsten.behrens@bfh.ch}
}
