% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Labdat.r
\name{ccCleanLabLoadDirect}
\alias{ccCleanLabLoadDirect}
\title{Extract and combine raw data from ccCleanLabLoadDirect}
\usage{
ccCleanLabLoadDirect(
  vsFieldCodes = "",
  sDbCleanLabVersion = "DbCleanLab",
  sConfigFile = "config.yml"
)
}
\arguments{
\item{sDbCleanLabVersion}{Version of DbCleanLab to use;  Default: "DbCleanLab". "DbCleanLabDevelopment" should be used for testing during development}

\item{sConfigFile}{Path to the config file (.yml) with the database credentials}

\item{vsIdSampeKobo}{vsIdSampeKobo Field codes to extract; vector of field codes. Names must be the same as in CleanLab column "field_code" (e.g. "PREZ 123")}
}
\value{
list of dataframes
}
\description{
Load various datasets from db based on project names
}
\author{
Daniel Wächter, \email{daniel.waechter@bfh.ch}
}
